Address comments and fix linter failure
